+++ src/components/admin/sections/FileSupport.vue
@@ -28,6 +28,15 @@
             : t('memories', 'Thumbnail generation may not work for videos.')
         }}
       </NcNoteCard>
+
+      <NcNoteCard v-if="hasProvider('OC\\Preview\\Imaginary')" type="warning">
+        {{
+          t(
+            'memories',
+            'Imaginary is enabled for preview generation. This will override other preview providers. We currently recommend against using Imaginary due to multiple bugs in handling of HEIC and GIF files.',
+          )
+        }}
+      </NcNoteCard>
     </template>
 
     <NcCheckboxRadioSwitch
@@ -110,6 +119,9 @@ export default defineComponent({
       },
       'OC\\Preview\\Movie': {
         name: t('memories', 'Videos (ffmpeg)'),
+      },
+      'OC\\Preview\\Imaginary': {
+        name: t('memories', 'Imaginary (not recommended)'),
       },
     },